refactor(channels): F4.5 first slice — 抽核心 trait + manager + relay 到 dasclaw_channels
按 ADR-152 §3 F4.5（verbatim port，遵循 ADR-129 §1.3）：

新 crate crates/dasclaw_channels 内容：
- channel.rs：Channel trait / IncomingMessage / OutgoingResponse / StatusUpdate / AttachmentKind 等
- manager.rs：ChannelManager
- relay/{mod,channel,client,webhook}.rs：relay 子系统 4 文件
- error.rs：ChannelError（从 desktop Error 抽出）

桌面端通过 re-export shim 保持 0 调用方改动：
- channels/mod.rs 重写为薄壳（pub use dasclaw_channels::*）
- error.rs 把 ChannelError 改为 pub use dasclaw_channels::ChannelError
- 新增 channels/status.rs::build_tool_completed 替代原 StatusUpdate::tool_completed
  inherent method（Rust 禁止外部类型 inherent impl），更新 5 个调用点
- 7 个 manager 单测搬到 tests/channels_manager.rs（用 public API health_check_all 取代私字段访问）

不在本 PR：wasm/web/repl/http/signal/webhook_server 留 desktop（重 GUI/系统资源），
属 F4.5.2/3 后续切片。

Closes #732
