+++ Cargo.toml
@@ -31,6 +31,8 @@ members = [
 	"crates/dasclaw_lsp",
 	"crates/dasclaw_git_tools",
 	"crates/dasclaw_routines",
+	# F4.5 first slice — channel core trait + manager + relay extracted from desktop
+	"crates/dasclaw_channels",
 	# W2.6 接线层：组合 sandbox + pty + workspace_cap policy
 	"crates/dasclaw_exec",
 	# W6-C 主仓自实现 LLM provider —— 替代 claw-code-api path-dep（ADR-118 PR-A.0 骨架）
