+++ crates/dasclaw_channels/src/channel.rs
@@ -96,7 +96,7 @@ pub struct IncomingMessage {
     /// Internal-only flag: message was generated inside the process (e.g. job
     /// monitor) and must bypass the normal user-input pipeline. This field is
     /// not settable via metadata, so external channels cannot spoof it.
-    pub(crate) is_internal: bool,
+    pub is_internal: bool,
 }
 
 impl IncomingMessage {
@@ -176,7 +176,7 @@ impl IncomingMessage {
     }
 
     /// Mark this message as internal (bypasses user-input pipeline).
-    pub(crate) fn into_internal(mut self) -> Self {
+    pub fn into_internal(mut self) -> Self {
         self.is_internal = true;
         self
     }
@@ -357,38 +357,6 @@ pub enum StatusUpdate {
     },
 }
 
-impl StatusUpdate {
-    /// Build a `ToolCompleted` status with redacted parameters.
-    ///
-    /// On failure, serializes the tool's input parameters as pretty JSON after
-    /// replacing any keys listed in the tool's `sensitive_params()` with
-    /// `"[REDACTED]"`. On success, no parameters or error are included.
-    ///
-    /// Pass the resolved `Tool` reference (if available) so this method can
-    /// query `sensitive_params()` directly — callers don't need to manage the
-    /// borrow lifetime of the sensitive slice.
-    pub fn tool_completed(
-        name: String,
-        result: &Result<String, crate::error::Error>,
-        params: &serde_json::Value,
-        tool: Option<&dyn crate::tools::Tool>,
-    ) -> Self {
-        let success = result.is_ok();
-        let sensitive = tool.map(|t| t.sensitive_params()).unwrap_or(&[]);
-        Self::ToolCompleted {
-            name,
-            success,
-            error: result.as_ref().err().map(|e| e.to_string()),
-            parameters: if !success {
-                let safe = crate::tools::redact_params(params, sensitive);
-                Some(serde_json::to_string_pretty(&safe).unwrap_or_else(|_| safe.to_string()))
-            } else {
-                None
-            },
-        }
-    }
-}
-
 /// Enrich channel metadata with tool-call information.
 ///
 /// Returns a **clone** of `base` with `_tool_call_id` and (optionally)
@@ -505,130 +473,6 @@ pub trait ChannelSecretUpdater: Send + Sync {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::testing::credentials::TEST_REDACT_SECRET_123;
-
-    /// Stub tool that marks `"value"` as sensitive.
-    struct SecretTool;
-
-    #[async_trait]
-    impl crate::tools::Tool for SecretTool {
-        fn name(&self) -> &str {
-            "secret_save"
-        }
-        fn description(&self) -> &str {
-            "stub"
-        }
-        fn parameters_schema(&self) -> serde_json::Value {
-            serde_json::json!({"type": "object", "properties": {}})
-        }
-        async fn execute(
-            &self,
-            _params: serde_json::Value,
-            _ctx: &mut dyn dasclaw_runtime::JobContextCore,
-        ) -> Result<crate::tools::ToolOutput, crate::tools::ToolError> {
-            unreachable!()
-        }
-        fn sensitive_params(&self) -> &[&str] {
-            &["value"]
-        }
-    }
-
-    #[test]
-    fn tool_completed_redacts_sensitive_params_on_failure() {
-        let params = serde_json::json!({"name": "api_key", "value": TEST_REDACT_SECRET_123});
-        let err: Result<String, crate::error::Error> =
-            Err(crate::error::ToolError::ExecutionFailed {
-                name: "secret_save".into(),
-                reason: "db error".into(),
-            }
-            .into());
-        let tool = SecretTool;
-
-        let status = StatusUpdate::tool_completed(
-            "secret_save".into(),
-            &err,
-            &params,
-            Some(&tool as &dyn crate::tools::Tool),
-        );
-
-        if let StatusUpdate::ToolCompleted {
-            success,
-            error,
-            parameters,
-            ..
-        } = &status
-        {
-            assert!(!success);
-            let err_msg = error.as_deref().expect("should have error");
-            assert!(err_msg.contains("db error"), "error: {}", err_msg);
-            let param_str = parameters
-                .as_ref()
-                .expect("should have parameters on failure");
-            assert!(
-                param_str.contains("[REDACTED]"),
-                "sensitive value should be redacted: {}",
-                param_str
-            );
-            assert!(
-                !param_str.contains(TEST_REDACT_SECRET_123),
-                "raw secret should not appear: {}",
-                param_str
-            );
-            assert!(
-                param_str.contains("api_key"),
-                "non-sensitive params should be preserved: {}",
-                param_str
-            );
-        } else {
-            panic!("expected ToolCompleted variant");
-        }
-    }
-
-    #[test]
-    fn tool_completed_no_params_on_success() {
-        let params = serde_json::json!({"name": "key", "value": "secret"});
-        let ok: Result<String, crate::error::Error> = Ok("done".into());
-
-        let status = StatusUpdate::tool_completed("secret_save".into(), &ok, &params, None);
-
-        if let StatusUpdate::ToolCompleted {
-            success,
-            error,
-            parameters,
-            ..
-        } = &status
-        {
-            assert!(success);
-            assert!(error.is_none());
-            assert!(parameters.is_none(), "no params should be sent on success");
-        } else {
-            panic!("expected ToolCompleted variant");
-        }
-    }
-
-    #[test]
-    fn tool_completed_no_tool_passes_params_unredacted() {
-        let params = serde_json::json!({"cmd": "ls -la"});
-        let err: Result<String, crate::error::Error> =
-            Err(crate::error::ToolError::ExecutionFailed {
-                name: "shell".into(),
-                reason: "timeout".into(),
-            }
-            .into());
-
-        let status = StatusUpdate::tool_completed("shell".into(), &err, &params, None);
-
-        if let StatusUpdate::ToolCompleted { parameters, .. } = &status {
-            let param_str = parameters.as_ref().expect("should have parameters");
-            assert!(
-                param_str.contains("ls -la"),
-                "non-sensitive params should pass through: {}",
-                param_str
-            );
-        } else {
-            panic!("expected ToolCompleted variant");
-        }
-    }
 
     #[test]
     fn test_incoming_message_with_timezone() {
