+++ desktop-client/ironclaw/Cargo.toml
@@ -137,6 +137,8 @@ dasclaw_apply_patch = { path = "../../crates/dasclaw_apply_patch", version = "0.
 dasclaw_workspace_cap = { path = "../../crates/dasclaw_workspace_cap", version = "0.1.0" }
 # ADR-152 §3 F4.2.0 — `routine` 类型 + `RoutineError` verbatim 搬迁
 dasclaw_routines = { path = "../../crates/dasclaw_routines", version = "0.0.0-w1" }
+# ADR-152 §3 F4.5 (first slice) — channel core trait + manager + relay extracted
+dasclaw_channels = { path = "../../crates/dasclaw_channels", version = "0.0.0-w1" }
 # ADR-149 / issue #485 — Tool Visibility Triple Gate (`ToolVisibilityPolicy`
 # trait + closed-enum decisions). Default-OFF in the crate, host opts in.
 dasclaw_governance = { path = "../../crates/dasclaw_governance", version = "0.0.0-w1", features = ["tool_visibility"] }
